Small improvements & reorganizations:
1) Add an extended .gitignore
2) Modify YEAR to 2019 in LICENSE file
3) utils.py -> folder with input_generator.py and parser.py files.
4) Move examples to click
