+++ aiida_cp2k/utils/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+###############################################################################
+# Copyright (c), The AiiDA-CP2K authors.                                      #
+# SPDX-License-Identifier: MIT                                                #
+# AiiDA-CP2K is hosted on GitHub at https://github.com/aiidateam/aiida-cp2k   #
+# For further information on the license, see the LICENSE.txt file.           #
+###############################################################################
+"""AiiDA-CP2K utils"""
+from __future__ import absolute_import
+
+from .input_generator import Cp2kInput
+from .parser import parse_cp2k_output
+from .parser import parse_cp2k_trajectory
