+++ examples/single_calculations/test_failure.py
@@ -11,6 +11,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 
 import sys
+import click
 
 from aiida.orm import (Code, Dict)
 from aiida.common import NotExistent
@@ -18,46 +19,47 @@ from aiida.engine import run
 from aiida.common.exceptions import OutputParsingError
 from aiida_cp2k.calculations import Cp2kCalculation
 
-# ==============================================================================
-if len(sys.argv) != 2:
-    print("Usage: test_failure.py <code_name>")
-    sys.exit(1)
 
-codename = sys.argv[1]
+@click.command('cli')
+@click.argument('codelabel')
+def main(codelabel):
+    """Run failing calculation"""
+    try:
+        code = Code.get_from_string(codelabel)
+    except NotExistent:
+        print("The code '{}' does not exist".format(codelabel))
+        sys.exit(1)
 
-try:
-    code = Code.get_from_string(codename)
-except NotExistent:
-    print("The code '{}' does not exist".format(codename))
-    sys.exit(1)
+    print("Testing CP2K failure...")
 
-print("Testing CP2K failure...")
-
-# a broken CP2K input
-parameters = Dict(dict={'GLOBAL': {'FOO_BAR_QUUX': 42}})
-options = {
-    "resources": {
-        "num_machines": 1,
-        "num_mpiprocs_per_machine": 1,
-    },
-    "max_wallclock_seconds": 1 * 2 * 60,
-}
-
-print("Submitted calculation...")
-inputs = {
-    'parameters': parameters,
-    'code': code,
-    'metadata': {
-        'options': options,
+    # a broken CP2K input
+    parameters = Dict(dict={'GLOBAL': {'FOO_BAR_QUUX': 42}})
+    options = {
+        "resources": {
+            "num_machines": 1,
+            "num_mpiprocs_per_machine": 1,
+        },
+        "max_wallclock_seconds": 1 * 2 * 60,
     }
-}
-try:
-    run(Cp2kCalculation, **inputs)
-    print("ERROR!")
-    print("CP2K failure was not recognized")
-    sys.exit(3)
-except OutputParsingError:
-    print("CP2K failure correctly recognized")
 
-sys.exit(0)
-# EOF
+    print("Submitted calculation...")
+    inputs = {
+        'parameters': parameters,
+        'code': code,
+        'metadata': {
+            'options': options,
+        }
+    }
+    try:
+        run(Cp2kCalculation, **inputs)
+        print("ERROR!")
+        print("CP2K failure was not recognized")
+        sys.exit(3)
+    except OutputParsingError:
+        print("CP2K failure correctly recognized")
+
+    sys.exit(0)
+
+
+if __name__ == '__main__':
+    main()  # pylint: disable=no-value-for-parameter
